switched back to openssl
Switch the configure scripts and sources back to use openssl instead of gnutsl.
It is more likely that openssl is installed in the system than gnutls. Fixed
some signed/unsigned and const issues. Also changed the compiler arguments to
build for minimized size and increased the compilers diagnostics to catch more
issues.